refactor: move `buildenv.cc` declarations to `buildenv.hh`
