+++ pkgdb/src/buildenv/buildenv.cc
@@ -8,7 +8,6 @@
  * -------------------------------------------------------------------------- */
 
 #include <algorithm>
-#include <fcntl.h>
 #include <sys/stat.h>
 #include <sys/types.h>
